feat( catchgateway ): fix delete host info by host_id when make_room duplicated
